scripts/requirements.txt
pandas
matplotlib
seaborn
google-genai
tqdm
python-dotenv
qdrant-client
sentence-transformers
openai